Tiny-cloud setup and configuration.
Basic confugraiton of Tiny-cloud, currently it only supports ssh-keys
and hostname. In the future it might be useful to check if it can be
used for static IPs rather then using DHCP.

An alterantiv would either be to setup cloud-init, or have a internal
DHCP Server that updates DNS records for new machines.
As all servers should be internally be reachable over DNS rather then
IPs only.